--- components/camel-aws/camel-aws2-cw/src/test/java/org/apache/camel/component/aws2/cw/integration/CwComponentIntegrationTest.java
@@ -1,61 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.component.aws2.cw.integration;
-
-import org.apache.camel.EndpointInject;
-import org.apache.camel.Exchange;
-import org.apache.camel.ExchangePattern;
-import org.apache.camel.Processor;
-import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.aws2.cw.Cw2Constants;
-import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.camel.test.junit5.CamelTestSupport;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-
-@Disabled("Must be manually tested. Provide your own accessKey and secretKey!")
-public class CwComponentIntegrationTest extends CamelTestSupport {
-
-    @EndpointInject("mock:result")
-    private MockEndpoint mock;
-
-    @Test
-    public void sendInOnly() throws Exception {
-        mock.expectedMessageCount(1);
-
-        template.send("direct:start", ExchangePattern.InOnly, new Processor() {
-            public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(Cw2Constants.METRIC_NAME, "ExchangesCompleted");
-                exchange.getIn().setHeader(Cw2Constants.METRIC_VALUE, "2.0");
-                exchange.getIn().setHeader(Cw2Constants.METRIC_UNIT, "Count");
-            }
-        });
-
-        assertMockEndpointsSatisfied();
-    }
-
-    @Override
-    protected RouteBuilder createRouteBuilder() throws Exception {
-        return new RouteBuilder() {
-            @Override
-            public void configure() throws Exception {
-                from("direct:start").to("aws2-cw://http://camel.apache.org/aws-cw?accessKey=XXX&secretKey=XXX")
-                        .to("mock:result");
-            }
-        };
-    }
-}
